+++ IWindow/IWindow.hpp
@@ -27,7 +27,7 @@ class IWindow
     public:
 
         virtual ResponseType getResponse() = 0;
-        virtual void draw(GameField const&, size_t score, size_t speed, GameMode) = 0;
+        virtual void draw(GameField const&, size_t score, double speed, GameMode) = 0;
         virtual void openWindow(size_t width, size_t height) = 0;
         virtual void closeWindow() = 0;
         virtual void showGameOver() = 0;
